Fix ffmpeg command
Revert some attempts to fix error in transcoding causing a `encoded 0 frames` for some images.

The issue seams to rather come when stdout is redirected to a file. Using stdout seams to resolve the issue including for ImageNet n09472597_4754.JPEG. That specific image was failing even with `-report`  set to on and `-loglevel` set to debug